First commit
Some considerations: this project is using component reuse strategy, so the scripts might not make sense at first (they are from another project we have done). It is still in progress, so many things will change from now on.

Some variables and class names are in portuguese, our native language.